frontend/backend/pyproject.toml
[project]
name = "copilot-scaffold-backend"
version = "0.1.0"
description = "CopilotKit + LangGraph + FastAPI backend scaffold"
requires-python = ">=3.12"
dependencies = [
    "copilotkit[langgraph]==0.1.75",
    "ag-ui-langgraph>=0.0.26",
    "langgraph>=0.3.25,<1.1.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=4.0.0",
    "langchain-openai>=0.3.0,<0.4.0",
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "python-dotenv>=1.0.0",
    # RAG / Supabase
    "supabase>=2.4.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    # Ingest pipeline
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
